requirements.txt
tensorflow==2.15.0
numpy
matplotlib
pandas
scikit-learn
gradio
Pillow
